% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish}
\alias{fish}
\title{Stockfish engine}
\description{
This class wraps \code{bigchess}'s functions for handling
UCI-compatible engines. Instead of a functional workflow, this OOP approach
allows for more persistent interations with the engine.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{process}}{Connection to \code{{processx}} process running engine}

\item{\code{output}}{String vector with the latest output of the engine}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fish$new()}}
\item \href{#method-cmd}{\code{fish$cmd()}}
\item \href{#method-go}{\code{fish$go()}}
\item \href{#method-isready}{\code{fish$isready()}}
\item \href{#method-position}{\code{fish$position()}}
\item \href{#method-quit}{\code{fish$quit()}}
\item \href{#method-stop}{\code{fish$stop()}}
\item \href{#method-uci}{\code{fish$uci()}}
\item \href{#method-ucinewgame}{\code{fish$ucinewgame()}}
\item \href{#method-clone}{\code{fish$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Start Stockfish engine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to Stockfish executable (defaults to bundled version)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cmd"></a>}}
\if{latex}{\out{\hypertarget{method-cmd}{}}}
\subsection{Method \code{cmd()}}{
Send command to Stockfish engine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$cmd(command)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{command}}{String with UCI command}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-go"></a>}}
\if{latex}{\out{\hypertarget{method-go}{}}}
\subsection{Method \code{go()}}{
Start calculating on the current position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$go(
  wtime = NULL,
  btime = NULL,
  winc = NULL,
  binc = NULL,
  depth = NULL,
  infinite = FALSE,
  time = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wtime}}{Time (in msec) white has left on the clock}

\item{\code{btime}}{Time (in msec) black has left on the clock}

\item{\code{winc}}{White increment (in msec) per move if \code{wtime > 0}}

\item{\code{binc}}{Black increment (in msec) per move if \code{btime > 0}}

\item{\code{depth}}{Maximum number of plies to search}

\item{\code{infinite}}{Whether to search until \code{stop()} command}

\item{\code{time}}{Timeout (in seconds) for infinite computation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isready"></a>}}
\if{latex}{\out{\hypertarget{method-isready}{}}}
\subsection{Method \code{isready()}}{
Wait for the engine to be ready again
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$isready()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-position"></a>}}
\if{latex}{\out{\hypertarget{method-position}{}}}
\subsection{Method \code{position()}}{
Set up the position on the internal board
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$position(moves = NULL, startpos = TRUE, fen = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moves}}{String with LAN moves}

\item{\code{startpos}}{Whether starting from start position}

\item{\code{fen}}{String with FEN board}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-quit"></a>}}
\if{latex}{\out{\hypertarget{method-quit}{}}}
\subsection{Method \code{quit()}}{
Kill Stockfish engine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$quit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop calculating as soon as possible
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uci"></a>}}
\if{latex}{\out{\hypertarget{method-uci}{}}}
\subsection{Method \code{uci()}}{
Tell engine to use the UCI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$uci()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ucinewgame"></a>}}
\if{latex}{\out{\hypertarget{method-ucinewgame}{}}}
\subsection{Method \code{ucinewgame()}}{
Tell the next search will be from a different game
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$ucinewgame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fish$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
