% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish_find}
\alias{fish_find}
\title{Find bundled Stockfish executable}
\usage{
fish_find()
}
\description{
This package comes bundled with Stockfish, an open source, powerful UCI chess
engine. For more information about what Stockfish is, see the full
documentation of this package by running \code{?fish}.
}
