% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish_install}
\alias{fish_install}
\title{Install Stockfish executable}
\usage{
fish_install(path = NULL)
}
\arguments{
\item{path}{Where to place Stockfish's executable (by default, a folder named
'r-stockfish' inside \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}})}
}
\description{
Find, build, and install the latest version of Stockfish, an open source,
powerful UCI chess engine. For more information about what Stockfish is, see
the full documentation of this package by running \code{?fish}.
}
